Move providers/ to endpoint/providers/
