+++ cmd/ads/ads.go
@@ -19,6 +19,9 @@ import (
 	"github.com/open-service-mesh/osm/pkg/constants"
 	"github.com/open-service-mesh/osm/pkg/debugger"
 	"github.com/open-service-mesh/osm/pkg/endpoint"
+	"github.com/open-service-mesh/osm/pkg/endpoint/providers/azure"
+	azureResource "github.com/open-service-mesh/osm/pkg/endpoint/providers/azure/kubernetes"
+	"github.com/open-service-mesh/osm/pkg/endpoint/providers/kube"
 	"github.com/open-service-mesh/osm/pkg/envoy/ads"
 	"github.com/open-service-mesh/osm/pkg/featureflags"
 	"github.com/open-service-mesh/osm/pkg/httpserver"
@@ -27,9 +30,6 @@ import (
 	"github.com/open-service-mesh/osm/pkg/logger"
 	"github.com/open-service-mesh/osm/pkg/metricsstore"
 	"github.com/open-service-mesh/osm/pkg/namespace"
-	"github.com/open-service-mesh/osm/pkg/providers/azure"
-	azureResource "github.com/open-service-mesh/osm/pkg/providers/azure/kubernetes"
-	"github.com/open-service-mesh/osm/pkg/providers/kube"
 	"github.com/open-service-mesh/osm/pkg/signals"
 	"github.com/open-service-mesh/osm/pkg/smi"
 	"github.com/open-service-mesh/osm/pkg/utils"
