+++ pkg/endpoint/providers/azure/client.go
