+++ pkg/endpoint/providers/azure/errors.go
