+++ pkg/endpoint/providers/azure/kubernetes/client.go
