+++ pkg/endpoint/providers/azure/kubernetes/types.go
