+++ pkg/endpoint/providers/azure/provider.go
