+++ pkg/endpoint/providers/azure/provider_test.go
