+++ pkg/endpoint/providers/azure/suite_test.go
