+++ pkg/endpoint/providers/azure/types.go
