+++ pkg/endpoint/providers/azure/vm.go
