+++ pkg/endpoint/providers/azure/vmss.go
