+++ pkg/endpoint/providers/kube/client_test.go
